requirements.txt
dotenv
google-genai
youtube-transcript-api
pydantic
fastapi
uvicorn
